// src/api/generated/api.ts
export interface ApplicationsPostRequest {
    'name': string;
    'email': string;
    'resumeLink': string;
    'coverLetter'?: string;
    'portfolioLink'?: string;
    'mobile'?: string;
}
// ...
export interface JobsPostRequest {
    'title': string;
    'department': string;
    'location': string;
    'type'?: string;
    'salary'?: string;
    'description': string;
    'responsibilities'?: Array<string>;
    'requirements'?: Array<string>;
    'isActive'?: boolean;
}
export interface PaymentsCreatePost201Response {
    'success'?: boolean;
    'data'?: PaymentsCreatePost201ResponseData;
}
export interface PaymentsCreatePost201ResponseData {
    'orderId'?: string;
    'amount'?: number;
    'currency'?: string;
    'paymentId'?: string;
}
export interface PaymentsCreatePostRequest {
    'amount': number;
    'serviceName': string;
    'email'?: string;
    'mobile'?: string;
    'firstName'?: string;
    'lastName'?: string;
}
export interface PaymentsSuccessPostRequest {
    'paymentId': string;
    'orderId'?: string;
    'signature'?: string;
}
export interface PaymentsVerifyPostRequest {
    'paymentId': string;
    'orderId': string;
    'signature': string;
}
export interface SolutionContactContactPostRequest {
    'name': string;
    'email': string;
    'message': string;
    'mobile'?: string;
}
// ...
export class ApplicationsApi extends BaseAPI {
    /**
     * 
     * @summary Submit a general job application
     * @param {ApplicationsPostRequest} applicationsPostRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public applicationsPost(applicationsPostRequest: ApplicationsPostRequest, options?: RawAxiosRequestConfig) {
        return ApplicationsApiFp(this.configuration).applicationsPost(applicationsPostRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Apply for a specific job
     * @param {string} jobId ID of the job to apply for
     * @param {ApplicationsPostRequest} applicationsPostRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public jobsJobIdApplyPost(jobId: string, applicationsPostRequest: ApplicationsPostRequest, options?: RawAxiosRequestConfig) {
        return ApplicationsApiFp(this.configuration).jobsJobIdApplyPost(jobId, applicationsPostRequest, options).then((request) => request(this.axios, this.basePath));
    }
}
// ...
export class ContactApi extends BaseAPI {
    /**
     * 
     * @summary Submit a contact form
     * @param {SolutionContactContactPostRequest} solutionContactContactPostRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public solutionContactContactPost(solutionContactContactPostRequest: SolutionContactContactPostRequest, options?: RawAxiosRequestConfig) {
        return ContactApiFp(this.configuration).solutionContactContactPost(solutionContactContactPostRequest, options).then((request) => request(this.axios, this.basePath));
    }
}
// ...
export class JobsApi extends BaseAPI {
    /**
     * 
     * @summary Get all active job listings
     * @param {boolean} [showAll] Show all jobs (Admin only)
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public jobsGet(showAll?: boolean, options?: RawAxiosRequestConfig) {
        return JobsApiFp(this.configuration).jobsGet(showAll, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Create a new job listing (Admin)
     * @param {JobsPostRequest} jobsPostRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public jobsPost(jobsPostRequest: JobsPostRequest, options?: RawAxiosRequestConfig) {
        return JobsApiFp(this.configuration).jobsPost(jobsPostRequest, options).then((request) => request(this.axios, this.basePath));
    }
}
// ...
export class PaymentsApi extends BaseAPI {
    /**
     * 
     * @summary Create a payment order
     * @param {PaymentsCreatePostRequest} paymentsCreatePostRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public paymentsCreatePost(paymentsCreatePostRequest: PaymentsCreatePostRequest, options?: RawAxiosRequestConfig) {
        return PaymentsApiFp(this.configuration).paymentsCreatePost(paymentsCreatePostRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Record a successful payment
     * @param {PaymentsSuccessPostRequest} paymentsSuccessPostRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public paymentsSuccessPost(paymentsSuccessPostRequest: PaymentsSuccessPostRequest, options?: RawAxiosRequestConfig) {
        return PaymentsApiFp(this.configuration).paymentsSuccessPost(paymentsSuccessPostRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Verify a payment signature
     * @param {PaymentsVerifyPostRequest} paymentsVerifyPostRequest 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     */
    public paymentsVerifyPost(paymentsVerifyPostRequest: PaymentsVerifyPostRequest, options?: RawAxiosRequestConfig) {
        return PaymentsApiFp(this.configuration).paymentsVerifyPost(paymentsVerifyPostRequest, options).then((request) => request(this.axios, this.basePath));
    }
}